change precision to int16

+++ subtractAbs.m
@@ -1,3 +1,3 @@
 function out = subtractAbs(a,b)
-    out = abs(int8(a) - int8(b)); 
+    out = abs(int16(a) - int16(b)); 
 end
